Removed Warnings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@21291 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 204e4da48946ba38f22e37a072be64c288456a1e

+++ application/windscada/kohonen/src/kohonen/WTMLearningFunction.java
@@ -315,7 +315,7 @@ public class WTMLearningFunction implements LearningFunctionModel {
         Iterator<Integer> it = neighboorhood.keySet().iterator();
         int neuronNr;
         while(it.hasNext()){
-            neuronNr = (Integer)it.next();
+            neuronNr = it.next();
             if(showComments){
                 System.out.println("Changing neuron number: " + (neuronNr +1));
             }
